expression-evaluator: Added compiler flags for warnings and set the standard to c99.

+++ expression-evaluator/src/io-handling/input.c
@@ -9,7 +9,7 @@ read_from_stdin(void)
     char *line = NULL;
     size_t n = 0;
 
-    ssize_t line_len = getline(&line, &n, stdin);
+    int line_len = getline(&line, &n, stdin);
 
     if (line_len == -1) return NULL;
